Add pkpass save/share for boarding pass scan
This allows a scanned boarding pass to be converted to a pkpass file
and then shared with any application that can make use of it.  Note
that because of the limitations of what the boarding pass barcode
contains (no flight time, no frequent flyer status etc), the exported
pkpass is somewhat incomplete (so the user will have to add
information to certain fields).

This functionality is most useful with PassAndroid which allows pass
editing so you can fill in the missing information.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>